moved the dataset prep functions together
