+++ tools/data_processing/__init__.py
@@ -2,5 +2,6 @@ from .red_pajamav2 import load_redpajama
 from .tokenize import tokenize_dataset
 from .io import read_jsonl
 from .grouping import group_dataset
-from .qmsum import load_qmsum_test, load_qmsum_train
+from .qmsum import load_qmsum_train
 from .narrativeqa import load_narrativeqa_test, load_narrativeqa_train_valid
+from .prep_funcs import prepare_qmsum_test_ppl
