+++ tools/evaluation/eval.py
@@ -210,14 +210,13 @@ def main():
         from tools.data_processing.red_pajamav2 import load_redpajama
         test_ds = load_redpajama(tokenizer=tokenizer, split='train[90%:]', history_size=args.test_length, block_size=block_size, streaming=args.streaming, trust_remote_code=True)
     elif args.task_name == 'qmsum':
-        from tools.data_processing.qmsum import load_qmsum_test
-        test_ds = load_qmsum_test(max_token_num=args.max_context_length, test_length=args.test_length, block_size=block_size, tokenizer=tokenizer, split='test', with_answer=True)
-    elif args.task_name == 'ioeddk/qmsum':
-        from tools.data_processing.qmsum import load_qmsum_test
-        test_ds = load_qmsum_test(max_token_num=args.max_context_length, test_length=args.test_length, block_size=block_size, tokenizer=tokenizer, source='huggingface')
+        from tools.data_processing.prep_funcs import prepare_qmsum_test_ppl
+        test_ds = load_dataset(path="THUDM/LongBench", name="qmsum", split='test', streaming=args.streaming, trust_remote_code=True)
+        test_ds = prepare_test(test_ds, prepare_qmsum_test_ppl, max_token_num=args.max_context_length, test_length=args.test_length, block_size=block_size, tokenizer=tokenizer, with_answer=True)
     elif args.task_name == 'musique':
-        from tools.data_processing.musique import load_musique_test
-        test_ds = load_musique_test(test_length=args.test_length, block_size=block_size, tokenizer=tokenizer, with_answer=True)
+        from tools.data_processing.prep_funcs import prepare_musique_test_ppl
+        test_ds = load_dataset(path="THUDM/LongBench", name="musique", split='test', streaming=args.streaming, trust_remote_code=True)
+        test_ds = prepare_test(test_ds, prepare_musique_test_ppl, max_token_num=args.max_context_length, test_length=args.test_length, block_size=block_size, tokenizer=tokenizer, with_answer=True)
     else:
         from tools.registry import VALID_TASK_NAMES
         raise NotImplementedError(f"Task {args.task_name} is not supported, please choose from {VALID_TASK_NAMES}")
